feat(#54): Entriies search view : reset button + fieldset
