+++ assets/js/react/components/homepage/advancedSearch/AdvancedSearch.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable no-nested-ternary */
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
@@ -5,20 +6,26 @@ import AppBar from '@material-ui/core/AppBar';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import styled from 'styled-components';
-
+import { Card, CardContent } from '@material-ui/core';
+import CircularProgress from '@material-ui/core/CircularProgress';
 import EntriesSearch from './EntriesSearch';
-import Translate from '../common/Translate';
+import Translate from '../../common/Translate';
+
+import { entityOptionForSelector } from '../../../helpers/Entity';
 
 const advancedSearchTypes = ['entries', 'grottos', 'massifs'];
 
 const styles = theme => ({
   root: {
-    backgroundColor: theme.palette.background.paper,
+    backgroundColor: theme.palette.primary.dark,
     flexGrow: 1,
     width: '100%',
   },
+  resultsContainer: {
+    margin: '24px',
+  },
   tabName: {
-    fontSize: '1.2rem',
+    fontSize: '1.4rem',
   },
   tabContainer: {
     backgroundColor: theme.palette.primary3Color,
@@ -43,7 +50,9 @@ class AdvancedSearch extends React.Component {
   };
 
   render() {
-    const { classes, startAdvancedsearch } = this.props;
+    const {
+      classes, isLoading, results, resetAdvancedSearch, startAdvancedsearch,
+    } = this.props;
     const { value } = this.state;
 
     return (
@@ -85,15 +94,42 @@ class AdvancedSearch extends React.Component {
             />
           </Tabs>
         </AppBar>
-        {value === 0 && (
-        <div className={classes.tabContainer}>
-          <EntriesSearch
-            startAdvancedsearch={startAdvancedsearch}
-            resourceType={advancedSearchTypes[0]}
-          />
-        </div>
+
+        {results ? (
+          <Card className={classes.resultsContainer}>
+            <CardContent>
+              {isLoading ? (
+                <CircularProgress />
+              ) : (
+                results.length > 0 ? (
+                  results.map(result => (
+                    <div>
+                      {entityOptionForSelector({
+                        ...result,
+                        label: result.name,
+                      })}
+                    </div>
+                  ))
+                ) : (
+                  <Translate>Aucun résultat</Translate>
+                )
+              )}
+            </CardContent>
+          </Card>
+        ) : (
+          ''
         )}
-        {value === 1 && <div className={classes.tabContainer}>Groups search content</div> }
+
+        {value === 0 && (
+          <div className={classes.tabContainer}>
+            <EntriesSearch
+              startAdvancedsearch={startAdvancedsearch}
+              resourceType={advancedSearchTypes[0]}
+              resetResults={resetAdvancedSearch}
+            />
+          </div>
+        )}
+        {value === 1 && <div className={classes.tabContainer}>Groups search content</div>}
         {value === 2 && <div className={classes.tabContainer}>Massifs search content</div>}
       </div>
     );
@@ -102,6 +138,9 @@ class AdvancedSearch extends React.Component {
 
 AdvancedSearch.propTypes = {
   classes: PropTypes.shape({}).isRequired,
+  isLoading: PropTypes.bool.isRequired,
+  results: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+  resetAdvancedSearch: PropTypes.func.isRequired,
   startAdvancedsearch: PropTypes.func.isRequired,
 };
